move note embed method to NotesCommands
