+++ src/main/java/de/nplay/moderationbot/notes/NotesCommands.java
@@ -1,14 +1,19 @@
 package de.nplay.moderationbot.notes;
 
 import com.github.kaktushose.jda.commands.annotations.interactions.*;
+import com.github.kaktushose.jda.commands.dispatching.events.ReplyableEvent;
 import com.github.kaktushose.jda.commands.dispatching.events.interactions.CommandEvent;
 import com.github.kaktushose.jda.commands.dispatching.events.interactions.ModalEvent;
+import com.github.kaktushose.jda.commands.embeds.Embed;
 import de.nplay.moderationbot.Helpers;
 import de.nplay.moderationbot.permissions.BotPermissions;
+import net.dv8tion.jda.api.JDA;
 import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.entities.UserSnowflake;
 import net.dv8tion.jda.api.interactions.commands.Command.Type;
 import net.dv8tion.jda.api.interactions.components.text.TextInputStyle;
+
+import java.util.List;
 
 import static com.github.kaktushose.jda.commands.i18n.I18n.entry;
 
@@ -18,6 +23,13 @@ public class NotesCommands {
 
     private Member target;
     private boolean ephemeral = false;
+
+    public static Embed notesEmbed(ReplyableEvent<?> event, JDA jda, UserSnowflake target, List<NotesService.Note> notes) {
+        var targetUsername = jda.retrieveUserById(target.getIdLong()).complete().getName();
+        var embed = event.embed("noteList").placeholders(entry("target", targetUsername));
+        notes.stream().map(it -> it.toField(jda)).forEach(embed.fields()::add);
+        return embed;
+    }
 
     @Command("notes create")
     public void createNote(CommandEvent event, Member target) {
@@ -39,7 +51,7 @@ public class NotesCommands {
     @Command("notes list")
     public void listNotes(CommandEvent event, Member target) {
         var notes = NotesService.getNotesFromUser(target.getIdLong());
-        event.with().embeds(Helpers.notesEmbed(event, event.getJDA(), target, notes)).reply();
+        event.with().embeds(notesEmbed(event, event.getJDA(), target, notes)).reply();
     }
 
     @Command("notes delete")
